move extract ir to utils file
