+++ examples/example_unrolling_service/benchmarks/BUILD
@@ -1,3 +1,4 @@
+# TODO: use regular expressions instead of writing each file name manually
 exports_files([
     "offsets1.c",
     "conv2d.c",
